refactor: LadderString, LineString을 LadderSequence 하나의 클래스로 병합
