+++ src/test/java/util/LadderSequenceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-public class LadderStringTest {
+public class LadderSequenceTest {
 
     @DisplayName("Ladder의 상태를 String으로 반환한다.")
     @Test
@@ -16,7 +16,7 @@ public class LadderStringTest {
         Ladder ladder = new Ladder(new Height(5));
         CustomGenerator customGenerator = new CustomGenerator(List.of(false, true, false, true, false));
         ladder.init(5, customGenerator);
-        Assertions.assertThat(LadderString.from(ladder))
+        Assertions.assertThat(LadderSequence.from(ladder))
                 .isEqualTo(List.of(
                         "     |-----|     |-----|     |",
                         "     |-----|     |-----|     |",
